+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 - Added a nature topic.
 - Added shopping center to the city topic.
+- Added quizzes for degrees of comparison and a topic with degrees of comparison.
 
 ## v0.0.11 - 2022-10-27